Update SOJ-19 branch (#22)
* fix(navbar): padding (#13)

* build(roles/perms): added Spatie for roles/perms (#15)

* build(roles/perms): added Spatie for roles/perms

Spatie is added to the project. Role and Permission classes are made, and seeders are available for both classes (roles is filled with admin, teamleader, and user).

* refactor(roles/perms): revision databaseseeders

Revised seeddata in RoleSeeder: subroles added.

Added Role- and PermissionSeeders to DatabaseSeeder.

* feat(migrations): removed redundant migrations

The two removed migrations caused migration errors.

* fix(roles/perms): fixed issue with role assignment (#17)

Previously, when roles were to be assigned, an exception would be thrown regarding the missing table named 'role_user'. This table should have been generated automatically, but wasn't. The problem is fixed by manually creating a migration containing this table.

---------

Co-authored-by: Linde Rus <130562052+linderus@users.noreply.github.com>
Co-authored-by: Delite Borst <58737124+Auwburn@users.noreply.github.com>