feat(accounts): addition of account-promotion page (#31)
* feat(accounts): added visual layout

This is the main visual layout for the account management page. It is not functional as of yet.

* feat(localization): add Dutch localization SOJ-91

Added a localization file for the accounts view. Title is commented out as it is currently unknown.

* feat(accounts): replaced strings with localization

All hardcoded strings are replaced with localization strings, except the email (due to it being an example).

* refactor(js): moved script to separate file

JS script placed in own file instead in HTML

* feat(localization): added more strings for modal

Important strings for modal.

* feat(accounts): added modal w/ functionality

Added clickable modal to the accounts page. Also wrapped HTML in body.

* feat(accounts): added controller

Added an empty controller for the accounts view. Also added route.

* feat(accounts): table iterates through accounts

Not in conjunction with the database yet! Fake DB used for data.

* feat(accounts): added more styling for readability

The previous styling wasn't clear and readable after adding 50+ accounts. Color changes are made and hover effects added.

* feat(accounts): edited users are displayed

Edited users are now displayed in the confirmation modal. No ability to actually save new roles yet.

* feat(accounts): added toast-styling

Toast appears when pressing the "save" button in the modal, but it isn't responsive to an actual save yet.

* fix(roles/perms): fixed issue with role assignment

Previously, when roles were to be assigned, an exception would be thrown regarding the missing table named 'role_user'. This table should have been generated automatically, but wasn't. The problem is fixed by manually creating a migration containing this table.

* feat(seeddata): added seeddata for users

Also contains small improvement by adding guard names to roles.

* feat(accounts): roles are retrieved correctly

Houdt nog geen rekening met het feit dat er meerdere rollen zijn, maar het is voor nu belangrijk dat de juiste rollen worden opgehaald.

* refactor(lang): moved roles to separate file

Moved roles to own file as there are problems regarding saving translated values in the DB.

* feat(accounts): ability to update user roles

The user roles can be updated. Currently, the entire page refreshes whenever an update is made.

* feat(accounts): teamleader is combined role

The teamleader role cannot be chosen seperately; it is combined with the named team_roles ('bevers'). Due to this, all named teams are identified with 'team_'. This choice was based on the fact that one cannot be a teamleader without team.

* feat(toast): toast disappears after 3 sec

Previously, the toast didn't disappear after appearing for the first time. With a JS script, the opacity of the toast is set; opacity is used for a vanishing animation in the future.

* refactor(accounts): removal old toast

Old toast was redundant.

* feat(accounts): warning when saving wo/ change

Warning toast is shown when trying to save when no accounts have any role changes.

* feat(accounts): translated names in summary

Summary showed DB value names at first; which wasn't always as clear + it wasn't translated. Now the summary shows readable and translated values.

* Revert "feat(accounts): translated names in summary"

This reverts commit 3a4cafe535db0eb102088ed2ace190f8f63941d8.

* fix(accounts): correct selects are displayed

If a user was assigned teamleader, the default role of user was displayed. Now, the teamleader role itself is filtered out before it is send to the view, so it cannot be selected anymore.

* feat(lang): addition routes lang file

Same file as Jeroen

* refactor(accounts): removal header

Header is already included in base

* feat(accounts): added browser tests

* feat(accounts): added darkmode

View was unreadable without darkmode. Also, other views include darkmode.

* feat(accounts): added responsiveness tests

* refactor(accounts): added try/catch in update

* feat(accounts): added feature test

* feat(accounts): added specific toast content

Replaced placeholder message with usable one.

* refactor(accounts): changed names of tests

Previously the name referred to the controller, but the controller wasn't the only thing being tested. Name now includes all of manage-accounts.

* fix(accounts): solved merge conflict

Two routes were present for the manage-accounts view. All routes regarding manage-accounts are placed within the admin-middleware route.

* feat(accounts): tests are executed as admin

* fix(roles): all users are assigned w/ user role

Users weren't displayed as they lacked roles. Each created user is assigned the user role in the seeder.

* feat(accounts): added label for select

Added for accessibility.

* feat(accounts): redirect to home w/ exception

An exception sends the user back to the home screen

* fix(acc-promoveren): test en view classes

---------

Co-authored-by: Linde Rus <lsy.rus@student.avans.nl>